requirements.txt
# requirements.txt — clean production version

fastapi[standard]
uvicorn
python-multipart

sqlalchemy
asyncpg
alembic
greenlet
httpx
google-genai
passlib
python-jose
bcrypt
pydantic-settings
pydantic
python-dotenv
sentry-sdk
logfire
aiosqlite
redis
psycopg2
werkzeug
pyjwt
pwdlib[argon2]